# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {},
        "license": {}
    },
    "paths": {
...
        "/api/v1/room_prototypes": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Room Prototype"
                ],
                "summary": "获取聊天室原型列表",
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"data\":{},\"msg\":\"ok\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Room Prototype"
                ],
                "summary": "新增聊天室原型",
                "parameters": [
                    {
                        "type": "string",
                        "description": "PrototypeName",
                        "name": "prototype_name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "PeopleLimit",
                        "name": "people_limit",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Friendship",
                        "name": "friendship",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"data\":{},\"msg\":\"ok\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/room_prototypes/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Room Prototype"
                ],
                "summary": "根据ID获取聊天室原型",
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"data\":{},\"msg\":\"ok\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Room Prototype"
                ],
                "summary": "删除聊天室原型",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Room Prototype Id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"data\":{},\"msg\":\"ok\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/rooms": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Room"
                ],
                "summary": "获取聊天室列表",
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"data\":{},\"msg\":\"ok\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Room"
                ],
                "summary": "新增聊天室",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "RoomType",
                        "name": "room_type",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "PeopleLimit",
                        "name": "people_limit",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "RoomCount",
                        "name": "room_count",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"data\":{},\"msg\":\"ok\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/rooms/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Room"
                ],
                "summary": "获取指定聊天室",
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"data\":{},\"msg\":\"ok\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Room"
                ],
                "summary": "修改聊天室",
                "parameters": [
                    {
                        "type": "string",
                        "description": "RoomName",
                        "name": "room_name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "RoomDesc",
                        "name": "room_desc",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "RoomType",
                        "name": "room_type",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "PeopleLimit",
                        "name": "people_limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"data\":{},\"msg\":\"ok\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Room"
                ],
                "summary": "删除聊天室",
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"data\":{},\"msg\":\"ok\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/api/v1/users/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "获取指定user",
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"data\":{},\"msg\":\"ok\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "修改user",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Username",
                        "name": "username",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Nickname",
                        "name": "nickname",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "password",
                        "name": "password",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"data\":{},\"msg\":\"ok\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "删除user",
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"data\":{},\"msg\":\"ok\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    }
}